fix: clean -Wall -Wextra -Wpedantic warnings across engine + tests (0 warnings, 131/131 ctest green)
- crdt.c: cap make_id %s width so CRDT_ID_LEN never overflows (real UB fix)
- settings.c: JVal* -> const JVal* (discared-qualifier bug), split guard clauses
- view_doc.c: strdup doc_text() result instead of aliasing const (real const-bug fix)
- history/eqnum: add missing <stdio.h> (implicit sprintf/free decls)
- xps.h/pdfform.h/snippet.h: fix nested /* */ comment pitfall
- wubuimage/imgops, wubulayout/layout, wubumodel/model, cell_read, exp, rast,
  sync: split misleading-indentation one-liners; drop unused vars
- wubusvg/rast: add <stdio.h> for sscanf; drop unused atof_def
- pkgmgr/treeview/minimap/mdpreview: safe bounded copies (no strncpy truncation)
- 14 test files: add missing <stdlib.h>/<unistd.h> for free/getpid/memmem/unlink
- CMakeLists (viewshot): add missing src/wuburtf include + link libs (was a hard
  build break for the viewshot target)

+++ tests/test_imageops.c
@@ -123,7 +123,7 @@ int main(void) {
             if (cr) {
                 size_t cw = ocr_image_width(cr), ch = ocr_image_height(cr);
                 /* 20px square + 8px pad => ~28px, must be < 100 */
-                CHECK(cw < W && ch < H, "autocrop shrinks image");
+                CHECK(cw < (size_t)W && ch < (size_t)H, "autocrop shrinks image");
                 CHECK(cw >= 20 && ch >= 20, "autocrop keeps content");
                 ocr_image_free(cr);
             }
